Add link in banner to page about the Governor
The banner on https://nj.gov/ includes the text "Governor Phil Murphy • Lt. Governor Tahesha Way" that links to https://nj.gov/governor/. Previously, this link was not reflected in the NJWDS banner component.

This commit updates the banner Governor text to link to https://nj.gov/governor/. It opens in a new tab to mirror the behavior of the https://nj.gov/subscribe/ link.